Refactor `parser` package
