+++ src/test/java/athleticli/ui/NutrientVerifierTest.java
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import athleticli.parser.NutrientVerifier;
+
 class NutrientVerifierTest {
 
     @Test
